fix compile error with flag

+++ sources/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: tchoquet <tchoquet@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/17 17:48:37 by tchoquet          #+#    #+#             */
-/*   Updated: 2023/05/18 17:51:54 by tchoquet         ###   ########.fr       */
+/*   Updated: 2023/05/21 15:33:30 by tchoquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*output_str;
-	int		i;
+	size_t	i;
 
 	if (s == NULL)
 		return (NULL);
